+++ BakeryVendors/Controllers/VendorsController.cs
@@ -39,7 +39,7 @@ namespace BakeryVendors.Controllers
     }
 
     [HttpPost("/vendors/{vendorId}/orders")]
-    public ActionResult Create(int VendorId, string orderTitle, string orderDescription, int orderPrice, string orderDate)
+    public ActionResult Create(int VendorId, string orderTitle, string orderDescription, float orderPrice, string orderDate)
     {
       Dictionary<string, object> model = new Dictionary<string, object>();
       Vendors vendor = Vendors.Find(VendorId);
